Move scheduler into a single pkg
